Core: .ini now having 'env' mode parsing $VARIABLES (thanks Dolik), fixed ReadInt64 (thanks Hans)
git-svn-id: svn://ultimatepp.org/upp/trunk@6443 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/parser.cpp
@@ -171,7 +171,7 @@ int64 CParser::ReadInt64() throw(Error)
 
 int64 CParser::ReadInt64(int64 min, int64 max) throw(Error)
 {
-	int64 n = ReadInt();
+	int64 n = ReadInt64();
 	if(n < min || n > max)
 		ThrowError("number is out of range");
 	return n;
